requirements.txt
pywin32
kafka-python
Tk
